+++ system/scripts/actors/army-sheet.js
@@ -225,7 +225,7 @@ export class ArmySheetL5r5e extends BaseSheetL5r5e {
      * @private
      */
     async _updateLinkedActorData(type, actor) {
-        if (!actor || actor.documentName !== "Actor" || !["character", "npc"].includes(actor.data?.type)) {
+        if (!actor || actor.documentName !== "Actor" || !actor.isCharacter) {
             console.warn("L5R5E | Wrong actor type", actor?.data?.type, actor);
             return;
         }